requirements.txt
numpy==1.26.4
braindecode
wfdb
pyyaml
loguru
wandb

pyunpack 
patool

typing
dirhash
lightning
pytorch_lightning
torchaudio
rich
pytorch_metric_learning
captum
mkdocstrings

boto3
h5py

mkdocs-material

xlrd
scipy

black
black[jupyter]
isort

ipywidgets
ipykernel

seaborn
pandas

npy-append-array
rarfile

pyedflib

openpyxl

psg_utils
wfdb